requirements.txt
Django==3.2.4
django-allauth==0.44.0
django_crispy_forms==1.12.0
django-markdownx==3.0.1
Pillow==8.2.0